fix(core): handle concurrent modification notification service error
